% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{edge_transform_normal}
\alias{edge_transform_normal}
\title{Edge Transform Normal}
\usage{
edge_transform_normal(i, edges_midpoint, edges)
}
\arguments{
\item{i}{iterator}

\item{edges_midpoint}{midpoint of edges}

\item{edges}{edges to process}
}
\description{
Edge Transform Normal
}
