% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{get_rem}
\alias{get_rem}
\title{Get Transects}
\usage{
get_rem(linestring, raster, distance, length)
}
\arguments{
\item{linestring}{A data.frame of sf_LINESTRING}

\item{raster}{A raster (digital elevation model (DEM))}

\item{distance}{A \code{numeric} indicating distance between transects}

\item{length}{A \code{numeric} indicating the length of the transect}
}
\value{
A sf_POINT data.frame with elevations along transects (\code{elevation_adj}) and trend line (\code{elevation_main}).
}
\description{
Get Transects
}
\note{
Be aware of your crs projection as this will affect the length and distance arguments.
}
\examples{
\dontrun{
pts = matrix(c(170800,172000, 5410500, 5410400), 2)
line = sf::st_as_sf(sf::st_sfc(sf::st_linestring(pts), crs = 32612))

ele <- elevatr::get_elev_raster(line,
                                z = 13,
                                prj = '+proj=utm +zone=12 +datum=WGS84 +units=m +no_defs')

rem <- get_rem(line, ele, distance = 100, length = 500)

ele_crop <- terra::crop(terra::rast(ele), terra::vect(sf::st_buffer(line, 200)))
terra::plot(ele_crop)
plot(rem$geometry)
}
}
