% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{get_transect_points}
\alias{get_transect_points}
\title{Get Points along Transects}
\usage{
get_transect_points(linestring, distance, length)
}
\arguments{
\item{linestring}{A data.frame of sf_LINESTRING}

\item{distance}{A \code{numeric} indicating distance between transects}

\item{length}{A \code{numeric} indicating the length of the transect}
}
\value{
A list of geos_geometry POINTS along transects
}
\description{
Get Points along Transects
}
\examples{
\dontrun{
pts = matrix(c(170800,172000, 5410500, 5410400), 2)
line = sf::st_as_sf(sf::st_sfc(sf::st_linestring(pts), crs = 32612))

points <- get_transects(line, distance = 100, length = 500)

ele_crop <- terra::crop(terra::rast(ele), terra::vect(sf::st_buffer(line, 200)))
terra::plot(ele_crop)
plot(points)
}
}
