% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{line_perpendicular_to_edge}
\alias{line_perpendicular_to_edge}
\title{Line Perpendicular to Edge}
\usage{
line_perpendicular_to_edge(
  point,
  which_edge,
  length = 1,
  edges_midpoint,
  edges
)
}
\arguments{
\item{point}{sf POINT object}

\item{which_edge}{which edge of \code{edges} to process}

\item{length}{length of scale_x and scale_y in affline scale}

\item{edges_midpoint}{midpoint of edges}

\item{edges}{edges to process}
}
\description{
Line Perpendicular to Edge
}
