% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{ps_and_ts}
\alias{ps_and_ts}
\title{ps and ts}
\usage{
ps_and_ts(edges, distance, length, line)
}
\arguments{
\item{edges}{edges to process}

\item{distance}{A \code{numeric} indicating distance between transects}

\item{length}{A \code{numeric} indicating the length of the transect}

\item{line}{A linestring}
}
\description{
ps and ts
}
