% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{pts_on_transects}
\alias{pts_on_transects}
\title{Points on transects}
\usage{
pts_on_transects(pts_and_transects, distance)
}
\arguments{
\item{pts_and_transects}{created with get_pts_and_transects}

\item{distance}{A \code{numeric} indicating distance between transects}
}
\value{
Lists of sf_POINT data.frames
}
\description{
Points on transects
}
