% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transects.R
\name{get_transects}
\alias{get_transects}
\title{Get Transects}
\usage{
get_transects(linestring, raster, distance, length)
}
\arguments{
\item{linestring}{A data.frame of sf_LINESTRING}

\item{raster}{A raster (digital elevation model (DEM))}

\item{distance}{A \code{numeric} indicating distance between transects}

\item{length}{A \code{numeric} indicating the length of the transect}
}
\value{
A sfc_POINT data.frame with elevations along the transect as well as the linestring.
}
\description{
Get Transects
}
\note{
Be aware of your crs projection as this will affect the length and distance arguments.
}
\examples{
\dontrun{
line <- mapedit::drawFeatures() \%>\%
sf::st_transform(32612)
ele <- elevatr::get_elev_raster(line, z = 13)
terra::crs(ele) <- '+proj=utm +zone=12 +datum=WGS84 +units=m +no_defs'

rem <- get_transects(line, ele, distance = 100, length = 500)

ele_crop <- terra::crop(ele, terra::vect(sf::st_buffer(line, 200)))
terra::plot(ele_crop)
plot(line$geometry, add = TRUE)
plot(rem$geometry, add = TRUE)
}

}
