% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem_raster}
\alias{rem_raster}
\title{REM Raster}
\usage{
rem_raster(rem, raster, fun = "mean", window = 3, ...)
}
\arguments{
\item{rem}{A previously created \code{get_rem()} object}

\item{raster}{Raster used to generate \code{get_rem()}}

\item{fun}{A function to use in \code{focal()} for NA's}

\item{window}{Window size for \code{focal()}}

\item{...}{Additional arguments for \code{focal()}}
}
\value{
A terra raster object.
}
\description{
REM Raster
}
\note{
The raw output of 'get_rem()' contains holes when converting to a raster. Thus, a
focal function is used to clean up the holes.
}
\examples{
\dontrun{
pts = matrix(c(170800,172000, 5410500, 5410400), 2)
line = sf::st_as_sf(sf::st_sfc(sf::st_linestring(pts), crs = 32612))

ele <- elevatr::get_elev_raster(line,
                                z = 13,
                                prj = '+proj=utm +zone=12 +datum=WGS84 +units=m +no_defs')

rem <- get_rem(line, ele, distance = 100, length = 500)

rem_rast <- rem_raster(rem, ele, fun = 'mean', window = 3, na.rm = TRUE)

}

}
